fix: integrate transaction pool with activityTab
